+++ connect/src/main/java/com/stripe/android/connect/webview/serialization/SetCollectMobileFinancialConnectionsResultPayloadJs.kt
@@ -19,10 +19,9 @@ internal data class SetCollectMobileFinancialConnectionsResultPayloadJs(
     companion object {
         fun from(
             id: String,
-            result: FinancialConnectionsSheetResult?
+            result: FinancialConnectionsSheetResult
         ): SetCollectMobileFinancialConnectionsResultPayloadJs {
             return when (result) {
-                null,
                 is FinancialConnectionsSheetResult.Canceled -> {
                     SetCollectMobileFinancialConnectionsResultPayloadJs(
                         id = id,