+++ paymentsheet/src/main/java/com/stripe/android/lpmfoundations/paymentmethod/definitions/CashAppPayDefinition.kt
@@ -27,7 +27,10 @@ internal object CashAppPayDefinition : PaymentMethodDefinition {
 }
 
 private object CashAppPayUiDefinitionFactory : UiDefinitionFactory.RequiresSharedDataSpec {
-    override fun createSupportedPaymentMethod(sharedDataSpec: SharedDataSpec) = SupportedPaymentMethod(
+    override fun createSupportedPaymentMethod(
+        metadata: PaymentMethodMetadata,
+        sharedDataSpec: SharedDataSpec,
+    ) = SupportedPaymentMethod(
         paymentMethodDefinition = CashAppPayDefinition,
         sharedDataSpec = sharedDataSpec,
         displayNameResource = R.string.stripe_paymentsheet_payment_method_cashapp,
@@ -45,6 +48,7 @@ private object CashAppPayUiDefinitionFactory : UiDefinitionFactory.RequiresShare
             emptyList()
         }
         return transformSpecToElements.transform(
+            metadata = metadata,
             specs = sharedDataSpec.fields + localLayoutSpecs
         )
     }