+++ paymentsheet/src/main/java/com/stripe/android/lpmfoundations/paymentmethod/definitions/SunbitDefinition.kt
@@ -26,7 +26,10 @@ internal object SunbitDefinition : PaymentMethodDefinition {
 }
 
 private object SunbitUiDefinitionFactory : UiDefinitionFactory.RequiresSharedDataSpec {
-    override fun createSupportedPaymentMethod(sharedDataSpec: SharedDataSpec) = SupportedPaymentMethod(
+    override fun createSupportedPaymentMethod(
+        metadata: PaymentMethodMetadata,
+        sharedDataSpec: SharedDataSpec,
+    ) = SupportedPaymentMethod(
         paymentMethodDefinition = SunbitDefinition,
         sharedDataSpec = sharedDataSpec,
         displayNameResource = R.string.stripe_paymentsheet_payment_method_sunbit,